model/requirements.txt
pandas>=2.1.0
numpy>=1.26.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
google-genai>=0.5.0
groq>=0.5.0
tavily-python>=0.3.0
vaderSentiment>=3.3.2
python-dotenv>=1.0.0
pymongo>=4.6.0
scikit-learn>=1.3.0
torch>=2.2.0
pytorch-lightning>=2.2.0
pytorch-forecasting==1.1.1
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pytz>=2024.1
yfinance>=0.2.0
